requirements.txt
requests
python-dateutil
pillow